Added periodic GUI Updating and Dynamic HTML Rendering
The current way this is written in the update-index.js file causes the browser to throw one error because we don't wait for the asynchronous HTTP Request to complete before we try to access the data from the response. It still works but maybe something to fix at some point.